Shift PlotUploader class to separate file.
